# swagger/eth-auth-server-swagger.yml
swagger: '2.0'
info:
  version: 0.1.0
  title:  Ethjournal Authentication server
  description: ''
host: 'localhost:8080'
basePath: /
tags:
  - name: Authentication
    description: 'Authenticate users'
paths:
    '/v1/auth':
      post:
        operationId: authenticateUser
        tags:
        - Authentication
        produces:
        - application/json
        consumes:
        - application/json
        parameters:
        - name: login
          in: body
          schema:
            $ref: '#/definitions/model-login'
        responses:
          '200':
            description: ''
            schema:
              $ref: '#/definitions/response-model-token'
          '401':
            $ref: '#/responses/response-unauthorized'
          '500':
            $ref: '#/responses/response-internal-server-error-500'
definitions:
    response-base:
      type: object
      properties:
        success:
          type: boolean
        resultCode:
          type: integer
          format: int32
        message:
          type: string
        errors:
          type: array
          items:
            type: string
        transactionId:
          type: string
        type:
          type: string
      discriminator: type
      required:
        - success
        - resultCode
    model-login:
      type: object
      properties:
        userName:
          type: string
        password:
          type: string
      required:
         - userName
         - password
    response-model-token:
      allOf:
      - $ref: '#/definitions/response-base'
      - type: object
        properties:
          tokenString:
            type: string
responses:
    response-internal-server-error-500:
      description: Internal Server error
      schema:
        $ref: '#/definitions/response-base'
    response-unauthorized:
        description: Authentication information is missing or invalid
        schema:
          $ref: '#/definitions/response-base' 
        headers:
          WWW_Authenticate:
            type: string